requirements.txt
streamlit==1.31.0
pandas==2.2.0
pillow==10.2.0
stripe==7.11.0
bcrypt==4.1.2
python-dotenv==1.0.1
pymongo==4.6.1
